+++ api/index.js
@@ -9,6 +9,7 @@ import pois from './pois';
 import rides from './rides';
 import jwt from './jwt';
 import feedback from './feedback';
+import drivers from './drivers';
 
 export default function (ctx, inject) {
   const api = {
@@ -23,6 +24,7 @@ export default function (ctx, inject) {
     rides: rides(ctx.$axios),
     jwt: jwt(ctx.$axios),
     feedback: feedback(ctx.$axios),
+    drivers: drivers(ctx.$axios),
   };
   ctx.$api = api;
   inject('api', api);